refactor(frontend): flatten _engine layout route — move ns routes up to _context level
