+++ frontend/src/routes/_context.tsx
@@ -71,6 +71,7 @@ function RouteComponent() {
 			<Outlet />
 			<ModalRenderer />
 			<Modals />
+			{!features.multitenancy && <EngineModals />}
 		</>
 	);
 }
@@ -93,6 +94,27 @@ function Modals() {
 								...old,
 								modal: undefined,
 							}),
+						});
+					}
+				},
+			}}
+		/>
+	);
+}
+
+function EngineModals() {
+	const navigate = useNavigate();
+	const search = Route.useSearch();
+	const CreateNamespaceDialog = useDialog.CreateNamespace.Dialog;
+	return (
+		<CreateNamespaceDialog
+			dialogProps={{
+				open: search.modal === "create-ns",
+				onOpenChange: (value) => {
+					if (!value) {
+						return navigate({
+							to: ".",
+							search: (old) => ({ ...old, modal: undefined }),
 						});
 					}
 				},
