--- frontend/src/routes/_context/_engine/ns.$namespace/settings.tsx
@@ -1,8 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { RouteComponent } from "../../_cloud/orgs.$organization/projects.$project/ns.$namespace/settings";
-
-export const Route = createFileRoute(
-	"/_context/_engine/ns/$namespace/settings",
-)({
-	component: RouteComponent,
-});
