+++ frontend/src/routes/_context/ns.$namespace/settings.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { RouteComponent } from "../_cloud/orgs.$organization/projects.$project/ns.$namespace/settings";
+
+export const Route = createFileRoute("/_context/ns/$namespace/settings")({
+	component: RouteComponent,
+});
